+++ .gitignore
@@ -1,6 +1,6 @@
 
-# Created by https://www.toptal.com/developers/gitignore/api/windows,java-web,eclipse
-# Edit at https://www.toptal.com/developers/gitignore?templates=windows,java-web,eclipse
+# Created by https://www.toptal.com/developers/gitignore/api/windows,eclipse,java
+# Edit at https://www.toptal.com/developers/gitignore?templates=windows,eclipse,java
 
 ### Eclipse ###
 .metadata
@@ -68,9 +68,30 @@ local.properties
 # Spring Boot Tooling
 .sts4-cache/
 
-### Java-Web ###
-## ignoring target file
-target/
+### Java ###
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
 
 ### Windows ###
 # Windows thumbnail cache files
@@ -98,4 +119,4 @@ $RECYCLE.BIN/
 # Windows shortcuts
 *.lnk
 
-# End of https://www.toptal.com/developers/gitignore/api/windows,java-web,eclipse
+# End of https://www.toptal.com/developers/gitignore/api/windows,eclipse,java
