+++ src/com/ondongne/teampjt/circle/action/ActionDataCircle.java
@@ -1,7 +1,9 @@
+// 작성자: 김나예
+
 package com.ondongne.teampjt.circle.action;
 
-public interface ActionDataCircle {
+public interface ActionDataCircle{
 	
-	public void executeQuery();
+	public void execute();
 	
 }
