--- src/com/ondongne/teampjt/circle/action/dataAction.java
@@ -1,7 +1,0 @@
-package com.ondongne.teampjt.circle.action;
-
-public interface dataAction{
-	
-	public void execute();
-	
-}
