--- src/com/ondongne/teampjt/circle/dao/mybatis_config.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE configuration
-
-PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
-
-"http://mybatis.org/dtd/mybatis-3-config.dtd">
-
-<configuration>
-
-<environments default="development">
-        <environment id="development">
-            <transactionManager type="JDBC" />
-            <dataSource type="POOLED">
-                <property name="driver" value="com.mysql.jdbc.Driver" />
-                <property name="url" value="jdbc:mysql://localhost:3306/ondongne?characterEncoding=UTF8&amp;serverTimezone=UTC" />
-                <property name="username" value="dongne" />
-                <property name="password" value="34811005" />
-            </dataSource>
-        </environment>
-    </environments>
-
-<mappers>
-	<mapper resource="com/ondongne/teampjt/circle/dao/mapper.xml" />
-</mappers>
-
-</configuration>
