+++ src/com/ondongne/teampjt/circle/dto/DataTransferCircle.java
@@ -1,5 +1,107 @@
+// 작성자: 김나예
+
 package com.ondongne.teampjt.circle.dto;
 
 public class DataTransferCircle {
-
+	private String email;
+	private String title;
+	private String pictures;
+	private String region;
+	private String zipcode;
+	private String address;
+	private String address_detail;
+	private String event_date;
+	private String end_date;
+	private String gender;
+	private int start_age;
+	private int end_age;
+	private String description;
+	private String tags;
+	
+	public String getEmail() {
+		return email;
+	}
+	public void setEmail(String email) {
+		this.email = email;
+	}
+	public String getTitle() {
+		return title;
+	}
+	public void setTitle(String title) {
+		this.title = title;
+	}
+	public String getPictures() {
+		return pictures;
+	}
+	public void setPictures(String pictures) {
+		this.pictures = pictures;
+	}
+	public String getRegion() {
+		return region;
+	}
+	public void setRegion(String region) {
+		this.region = region;
+	}
+	public String getZipcode() {
+		return zipcode;
+	}
+	public void setZipcode(String zipcode) {
+		this.zipcode = zipcode;
+	}
+	public String getAddress() {
+		return address;
+	}
+	public void setAddress(String address) {
+		this.address = address;
+	}
+	public String getAddress_detail() {
+		return address_detail;
+	}
+	public void setAddress_detail(String address_detail) {
+		this.address_detail = address_detail;
+	}
+	public String getEvent_date() {
+		return event_date;
+	}
+	public void setEvent_date(String event_date) {
+		this.event_date = event_date;
+	}
+	public String getEnd_date() {
+		return end_date;
+	}
+	public void setEnd_date(String end_date) {
+		this.end_date = end_date;
+	}
+	public String getGender() {
+		return gender;
+	}
+	public void setGender(String gender) {
+		this.gender = gender;
+	}
+	public int getStart_age() {
+		return start_age;
+	}
+	public void setStart_age(int start_age) {
+		this.start_age = start_age;
+	}
+	public int getEnd_age() {
+		return end_age;
+	}
+	public void setEnd_age(int end_age) {
+		this.end_age = end_age;
+	}
+	public String getDescription() {
+		return description;
+	}
+	public void setDescription(String description) {
+		this.description = description;
+	}
+	public String getTags() {
+		return tags;
+	}
+	public void setTags(String tags) {
+		this.tags = tags;
+	}
+	
+	
 }
