+++ src/com/ondongne/teampjt/users/mapper.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
+
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="signup">
@@ -7,11 +8,11 @@
 		insert into users(email, password, joindate)
 		values(#{email}, #{password}, now())
 	</insert>
-	
+
 	<select id="checkEmail" resultType="String">
 		select email from users where email=#{email}
 	</select>
-	
+
 	<select id="checkSignin" resultType="com.ondongne.teampjt.users.dto.DataTransferSignup">
 		select email, password from users where email=#{email}
 	</select>
